exposes resident compiler
